Moved cardmanager to its own package
