+++ tradingcards-plugin/src/main/java/net/tinetwork/tradingcards/tradingcardsplugin/managers/cards/CompositeCardKey.java
@@ -1,4 +1,4 @@
-package net.tinetwork.tradingcards.tradingcardsplugin.managers;
+package net.tinetwork.tradingcards.tradingcardsplugin.managers.cards;
 
 import org.jetbrains.annotations.Contract;
 import org.jetbrains.annotations.NotNull;
